requirements.txt
fastchat-mcp==1.2.5
PyYAML